+++ packages/react/Checkbox/src/components/CheckboxSkeleton/index.ts
@@ -1,0 +1,4 @@
+import { CheckboxSkeleton } from "./CheckboxSkeleton";
+
+export { CheckboxSkeleton } from "./CheckboxSkeleton";
+export default CheckboxSkeleton;