+++ packages/styles/src/index.ts
@@ -1,7 +1,8 @@
 export { utils } from "./utils";
 
-export { chip } from "./packages/chip";
 export { badge } from "./packages/badge";
+export { checkbox } from "./packages/checkbox";
+export { chip } from "./packages/chip";
 export { icon } from "./packages/icon";
 export { label } from "./packages/label";
 export { skeleton } from "./packages/skeleton";