+++ packages/tokens/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Our design tokens are the foundations of Nimbus Design System.
 
+## 2022-09-24
+
+### 🎉 New features
+
+- Added new `text-disabled` and `surface-disabled` colors to the neutral color palette. ([#28](https://github.com/TiendaNube/nimbus-design-system/pull/28) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2022-09-01
 
 ### 🐛 Bug fixes